integrate leaflet for map

+++ airbnb-clone-back/src/main/java/fr/codecake/airbnb_clone_back/user/repository/UserRepository.java
@@ -1,13 +1,12 @@
 package fr.codecake.airbnb_clone_back.user.repository;
 
+import fr.codecake.airbnb_clone_back.user.domain.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
 import java.util.Optional;
 import java.util.UUID;
 
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import fr.codecake.airbnb_clone_back.user.domain.User;
-
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
     Optional<User> findOneByEmail(String email);
